Created bare-bones fetching structure
Overview
========

I created an index.html file in the root to serve as an example /
development page. (I just wanted to get started - we can absolutely come
up with a better place for this.)

I also moved widget.js to jquery.liveblog.js. After some thought, I'd
suggest simplifying the name of our plugin to just "liveblog",
especially since, as Jeremy and I discussed yesterday, we want the logic
that's hitting the API and returning data to be decoupled from any UI /
widget-specific logic.

As such, the plugin focuses purely on fetching updates and then
triggering "update" events on the jQuery object it's bound to that
passes along the relevant update data. I've created a very simple bit of
UI logic to push text updates to the page just for demonstration
purposes.