Added a pause method to $.liveBlogLiteApi
The API that repeatedly fetches updates from the LiveBlog URL can now be
paused by passing 'pause' as an argument to the plugin.

Ex: $('#myLiveBlog').liveBlogLiteApi('pause');

It stops all the calls to the internal fetch method until it's called
again. Because it keeps the same last_update key, whenever you unpause
it you'll get all the updates since you paused it.